test: atualiza teste do DynamicInputList e testes de snapshot
